Externalised packages

+++ install.sh
@@ -13,8 +13,13 @@ fi
 
 # Applications
 sudo apt update
-sudo apt upgrade
-sudo apt install nodejs npm
+sudo apt upgrade --yes
+wget -qO - https://adoptopenjdk.jfrog.io/adoptopenjdk/api/gpg/key/public | sudo apt-key add -
+sudo add-apt-repository --yes https://adoptopenjdk.jfrog.io/adoptopenjdk/deb/
+sudo apt install --yes $(awk '{print $1'} linux.txt)
 
 # Dotfiles
-ln -sv ~/dotfiles/git/.gitconfig ~
+if [[ ! -f ~/.gitconfig ]]
+then
+    ln -sv ~/dotfiles/git/.gitconfig ~
+fi
